pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "google-ads-mcp"
version = "0.1.0"
description = "MCP server for Google Ads Keyword Planning"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.2.0",
    "google-ads>=25.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
google-ads-mcp = "google_ads_mcp.server:main"
